Refactor personal dashboard functional test
